Add nbinteract back to gitbook
Although we won't know whether this will cause people to make a bunch of
requests to Binder again, since the banning code works correctly now we
should be able to do this without overloading the Binder servers. We can
then monitor the Binder requests to see if the textbook overloads the
Binder servers again.